chore(tests): 🔧 Consolidate pytest-html configuration into `conftest.py`
* Moved pytest-html configuration from `conftest_html.py` to `conftest.py`.
* Enhanced test reporting with additional metadata and failure details.
* Removed redundant `conftest_html.py` file to streamline test setup.
